Create new configuration for floodlight-docker interoperation
In this commit there are main changes, since the previous project
has been taken as a base to start our work. Indeed, most of the
previous work has gone, but it was a useful base to start meddling
with ovs.
Basically, there are three main modifications:
- Instead of using a local machine, Vagrant now needs an AWS
account, where it will create a EC2 instance with CentOS installed
and a floodlight image already downloaded
- Fix outdated nginx image (that now doesn't include `ping`
anymore) creating a custom one with the missing commands
- Add a script in /usr/bin, called `netboot`, that sets the
required network configurations and allows to easily try floodlight
with docker